Addressed more code review feedback
