+++ pytorch_lightning/cluster_environments/cluster_environment.py
@@ -11,12 +11,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Optional
 
-from pytorch_lightning.utilities import AMPType
+from pytorch_lightning.plugins.plugin import Plugin
 
 
-class ClusterEnvironment:
+class ClusterEnvironment(Plugin):
 
     def __init__(self):
         self._world_size = None
@@ -32,24 +31,3 @@ class ClusterEnvironment:
 
     def local_rank(self):
         pass
-
-    def required_plugins(self, amp_backend: AMPType) -> list:
-        """
-        Override to define additional required plugins. This is useful for when custom plugins
-        need to enforce override of other plugins.
-
-        Returns: Optional list of plugins containing additional plugins.
-
-        Example::
-            class MyPlugin(DDPPlugin):
-                def required_plugins(self):
-                    return [MyCustomAMPPlugin()]
-
-            # Will automatically add the necessary AMP plugin
-            trainer = Trainer(plugins=[MyPlugin()])
-
-            # Crash as MyPlugin enforces custom AMP plugin
-            trainer = Trainer(plugins=[MyPlugin(), NativeAMPPlugin()])
-
-        """
-        return []
