+++ pytorch_lightning/plugins/ddp_plugin.py
@@ -6,10 +6,10 @@ from pytorch_lightning import _logger as log
 from torch.optim import Optimizer
 from pytorch_lightning.core.lightning import LightningModule
 from pytorch_lightning.overrides.data_parallel import LightningDistributedDataParallel
-from pytorch_lightning.utilities import AMPType
+from pytorch_lightning.plugins.plugin import Plugin
 
 
-class DDPPlugin(object):
+class DDPPlugin(Plugin):
     """
     Plugin to link a custom ddp implementation to any arbitrary accelerator.
 
@@ -32,7 +32,7 @@ class DDPPlugin(object):
         self._ddp_kwargs: Dict[str, Any] = kwargs
 
     def configure_ddp(
-        self, model: LightningModule, device_ids: List[int]
+            self, model: LightningModule, device_ids: List[int]
     ) -> LightningDistributedDataParallel:
         """
         Pass through all customizations from constructor to `LightningDistributedDataParallel`.
@@ -69,12 +69,12 @@ class DDPPlugin(object):
         return model
 
     def init_ddp_connection(
-        self,
-        trainer,
-        cluster_environment,
-        global_rank: int,
-        world_size: int,
-        is_slurm_managing_tasks: bool = True,
+            self,
+            trainer,
+            cluster_environment,
+            global_rank: int,
+            world_size: int,
+            is_slurm_managing_tasks: bool = True,
     ) -> None:
         os.environ["MASTER_ADDR"] = str(cluster_environment.master_address())
         os.environ["MASTER_PORT"] = str(cluster_environment.master_port())
@@ -131,24 +131,3 @@ class DDPPlugin(object):
         if isinstance(model, LightningDistributedDataParallel):
             return model.module
         return model
-
-    def required_plugins(self, amp_backend: AMPType) -> list:
-        """
-            Override to define additional required plugins. This is useful for when custom plugins
-            need to enforce override of other plugins.
-
-        Returns: Optional list of plugins containing additional plugins.
-
-        Example::
-            class MyPlugin(DDPPlugin):
-                def required_plugins(self):
-                    return [MyCustomAMPPlugin()]
-
-            # Will automatically add the necessary AMP plugin
-            trainer = Trainer(plugins=[MyPlugin()])
-
-            # Crash as MyPlugin enforces custom AMP plugin
-            trainer = Trainer(plugins=[MyPlugin(), NativeAMPPlugin()])
-
-        """
-        return []
